requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
email-validator==2.1.1
psycopg2-binary==2.9.11
redis==7.1.0
argon2-cffi==25.1.0
PyJWT==2.10.1
gunicorn==23.0.0
concurrent-log-handler==0.9.28
